scripts/requirements.txt
asn1crypto==0.22.0
astroid==2.1.0
attrs==18.1.0
Automat==0.7.0
autopep8==1.4.3
cffi==1.10.0
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
enum34==1.1.6
ffmpeg-python==0.1.17
fire==0.1.3
future==0.17.1
hyperlink==18.0.0
idna==2.5
incremental==17.5.0
ipaddress==1.0.18
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.2.4
mccabe==0.6.1
numpy==1.16.1
opencv-python==4.0.0.21
packaging==16.8
parsel==1.5.0
pep8==1.7.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.5.0
pycparser==2.17
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pylint==2.2.2
pyparsing==2.2.0
queuelib==1.5.0
requests==2.13.0
Scrapy==1.5.0
service-identity==17.0.0
six==1.10.0
tqdm==4.11.2
Twisted==18.7.0
urllib3==1.21.1
virtualenv==16.4.0
w3lib==1.19.0
wrapt==1.11.1
